pyproject.toml
[tool.poetry]
name = "industry-news"
version = "0.1.0"
description = "AI, tech and startup news digest"
authors = ["Jakub Rakoczy <office@rakoczy.io>"]
license = "GNU GPLv3"
readme = "README.md"
include = ["resources/**/*"]

[tool.poetry.scripts]
run-main = "industry_news.__main__:main"

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.9.3"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.2.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
beautifulsoup4 = "4.12.3"
bs4 = "0.0.2"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
dataclasses-json = "0.6.4"
exceptiongroup = "1.2.0"
frozenlist = "1.4.1"
greenlet = "3.0.3"
idna = "3.6"
jsonpatch = "1.33"
jsonpointer = "2.4"
marshmallow = "3.20.2"
multidict = "6.0.5"
mypy = "1.8.0"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
packaging = "23.2"
pydantic = "2.6.1"
pydantic-core = "2.16.2"
pyyaml = "6.0.1"
requests = "2.31.0"
sniffio = "1.3.0"
soupsieve = "2.5"
sqlalchemy = "2.0.26"
tenacity = "8.2.3"
tomli = "2.0.1"
types-beautifulsoup4 = "4.12.0.20240106"
types-html5lib = "1.1.11.20240106"
types-requests = "2.31.0.20240125"
typing-inspect = "0.9.0"
typing-extensions = "4.9.0"
urllib3 = "2.2.0"
yarl = "1.9.4"
redditwarp = "^1.2.0"
furl = "^2.1.3"
langchain = "^0.1.12"
langchain-text-splitters = "^0.0.1"
tiktoken = "^0.6.0"
langchain-openai = "^0.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
